backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyjwt==2.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
email-validator==2.1.0
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0
groq==0.4.2
sqlalchemy-utils==0.41.1
sqlalchemy-json==0.7.0
fastapi-cors==0.0.6
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
